+++ src/pxlNav/effects/particles/SnowFall.js
@@ -8,7 +8,7 @@ import {
 } from "../../../libs/three/three.module.min.js";
 
 import ParticleBase from './ParticleBase.js';
-import { snowFallVert, snowFallFrag } from './shaders/SnowShader.js';
+import { snowSettings, snowFallVert, snowFallFrag } from './shaders/SnowShader.js';
 
 // Free floaties in the environment
 //   Dust balls & flakes
@@ -57,8 +57,10 @@ export class SnowFall extends ParticleBase{
       lightPos:{value:lightPosArr}
     };
         //let mtl = this.pxlFile.pxlShaderBuilder( snowUniforms, snowFallFrag( true ), snowFallFrag() );
+    
 
-    let mtl = this.room.pxlFile.pxlShaderBuilder( dustUniforms, snowFallVert( false, proxDist ), snowFallFrag() );
+    let curSnowSettings = Object.assign( {}, snowSettings, { "ProxDist": proxDist, "CalculateCollisions": true } );
+    let mtl = this.room.pxlFile.pxlShaderBuilder( dustUniforms, snowFallVert( curSnowSettings ), snowFallFrag() );
     mtl.side=DoubleSide;
     mtl.transparent=true;
     // mtl.blending=AdditiveBlending;
@@ -74,58 +76,3 @@ export class SnowFall extends ParticleBase{
     super.addToScene( vertexCount, pScale, mtl, 4, atlasPicks, randomAtlas );
   }
 }
-
-
-
-/*
-buildSnow(){
-  //sprite = ImageUtils.loadTexture( "textures/sprites/disc.png" );
-
-  let vertexCount = 12000; // Point Count
-  let pScale = 12;  // Point Base Scale
-
-  let geo = new BufferGeometry();
-  let verts = [];
-  let seeds = [];
-  let atlasId = [];
-
-  const atlasGen=()=>{ return Math.floor( (Math.random() * 4000) % 4 )*.25; };
-
-  for( let x=0; x<vertexCount; ++x ){
-    verts.push( 0,0,0 );
-    seeds.push( (Math.random()),(Math.random()),(Math.random()*2-1), (Math.random()*.5+.5) );
-    atlasId.push( atlasGen(), atlasGen() );
-  }
-
-  let posAttribute = new Float32BufferAttribute( verts, 3 );
-  let seedAttribute = new Float32BufferAttribute( seeds, 4 );
-  let atlasAttribute = new Float32BufferAttribute( atlasId, 2 );
-  //let idAttribute = new Uint8BufferAttribute( pId, 1 ); // ## would only be 0-65536; set up vector array for ids
-  geo.setAttribute( 'position', posAttribute );
-  geo.setAttribute( 'seeds', seedAttribute );
-  geo.setAttribute( 'atlas', atlasAttribute );
-  //geo.setAttribute( 'id', idAttribute );
-      
-  let snowUniforms={
-    snowTexture:{type:"t",value: this.pxlUtils.loadTexture( this.pxlUtils.assetRoot+"snow.jpg", 1, {"encoding":LinearSRGBColorSpace, "magFilter":NearestFilter, "minFilter":NearestMipmapNearestFilter} ) },
-    pointScale:{type:"f",value: pScale*this.pxlQuality.screenResPerc },
-    intensity:{type:"f",value:1.0},
-    rate:{type:"f",value:.035},
-  };
-  
-  let mtl = this.pxlFile.pxlShaderBuilder( snowUniforms, this.pxlShaders.particles.snowFallVert( this.mobile ), this.pxlShaders.particles.snowFallFrag() );
-  mtl.side=DoubleSide;
-  mtl.transparent=true;
-  mtl.blending=AdditiveBlending;
-  mtl.depthTest=true;
-  mtl.depthWrite=false;
-
-  let snow = new Points( geo, mtl );
-  snow.sortParticles = false;
-  snow.frustumCulled = false;
-  this.scene.add( snow );
-  snow.layers.set(1);
-  snow.pBaseScale=pScale;
-  this.geoList['snow']=snow;
-}
-  */