+++ src/pxlNav/effects/particles/shaders.js
@@ -2,11 +2,17 @@
 //   Kevin Edzenga, 2024
 
 //import { countDownVert, countDownFrag  } from './CountDownShader.js';
-import { emberWispsVert, emberWispsFrag } from './shaders/EmberWisps.js';
-import { dustVert, dustFrag } from './shaders/FloatingDust.js';
-import { heightMapVert, heightMapFrag } from './shaders/HeightMap.js';
-import { smokeVert, smokeFrag } from './shaders/Smoke.js';
+import { emberWispsSettings, emberWispsVert, emberWispsFrag } from './shaders/EmberWisps.js';
+import { dustSettings, dustVert, dustFrag } from './shaders/FloatingDust.js';
+import { heightMapSettings, heightMapVert, heightMapFrag } from './shaders/HeightMap.js';
+import { smokeSettings, smokeVert, smokeFrag } from './shaders/Smoke.js';
 //import { snowConfettiVert, snowConfettiFrag  } from './SnowConfettiShader.js';
-import { snowFallVert, snowFallFrag } from './shaders/SnowShader.js';
+import { snowSettings, snowFallVert, snowFallFrag } from './shaders/SnowShader.js';
 
-export { emberWispsVert, emberWispsFrag, dustVert, dustFrag, heightMapVert, heightMapFrag, smokeVert, smokeFrag, snowFallVert, snowFallFrag };
+export { 
+  emberWispsSettings, emberWispsVert, emberWispsFrag, 
+  dustSettings, dustVert, dustFrag,
+  heightMapSettings, heightMapVert, heightMapFrag,
+  smokeSettings, smokeVert, smokeFrag,
+  snowSettings, snowFallVert, snowFallFrag
+};